Split Controller's job listing and filtering
The splitting improves readability of the code
and lowers cyclomatic complexity of the functions.

Method JobsControllerImpl.List has been split into:
* filter - returning JobInfos matching filter;
* sort - sorting them;
* paginate - calculating returned page indexes and sizes.

The List function by itself stays responsible for setting up critical
section for accessing jobs collection, runs above functions and returns
proper records and info.

One mechanism requires clarification. If JobID used as index in pagination
does not match filter, it must be temporary added to Jobs collection
for calculation of the index position after sorting.
The "extra" variable is used for containing information, if this job
is added to the collection.

Method Job.passesFilter has been split into several methods.
Each of them verifies if one of the Job fields passes through the filter:
* Job.passesCreatedAfterFilter,
* Job.passesCreatedBeforeFilter,
* Job.passesUpdatedAfterFilter,
* Job.passesUpdatedBeforeFilter,
* Job.passesInfoFilter,
* Job.passesJobIDFilter,
* Job.passesNameFilter,
* Job.passesStatusFilter,
exactly as the name of the function suggests.

Change-Id: I03bab42db51c168afc06d5e954dfcbd137637f12
Signed-off-by: Lukasz Wojciechowski <l.wojciechow@partner.samsung.com>