Adjust validation to allow empty values
After user input is marshalled, swagger server validates data. When
input data is enum, it was strictly validated against enum values.
There is no empty value thus user receives an error and is forced to
always set enum fields in e.g. JobFilter.Status.

Modifying generated code is a bad practice but it was not possible (to
the best of the authors knowledge) to achieve this by simply modyfing
swagger.yml and re-generating models.

Possible solutions:
- adding and using new type (which would satisfy swaggers interfaces)
- customizing code generation

Change-Id: Ifc381a12936277db9938032dadb31fe865e84ad1
Signed-off-by: Alexander Mazuruk <a.mazuruk@samsung.com>