Fix descriptions in swagger.yml
Descriptions were not clear and/or did not document functionality
accurately enough. As docs are served with Weles it is important for the
API reference to be helpful.

While swagger-editor complained about adding descriptions to "$ref"
fields (as the referenced fields already had descriptions and should be
inherited by the referee), go-swagger did not add comments to structs
generated from definition with "$refs". This commit fixes that for
improved godoc documentation.

Additionally fixed swagger.yml style.

Change-Id: I28d1d8fbf73d043a21bf6473648d751df239b166
Signed-off-by: Alexander Mazuruk <a.mazuruk@samsung.com>