Fix mocks generation by Makefile
After moving to github.com and reorganizing mocks, the Makefile recipe
for mocks generation is no longer valid. It misses few folders causing
 manager/dryad/mock/session_provider.go
 manager/mock/communication_provider.go
 manager/mock/dryad_job_runner.go
not to be regenerated.

This patch replaces explicit list of folders to run go generate in with
generic ./... causing recursive regeneration of the whole tree.
Additionally a "-x" option is added for printing executed commands.

Change-Id: Ie35667daf1e1458c4bde63ffd0af0ee284628e95
Signed-off-by: Lukasz Wojciechowski <l.wojciechow@partner.samsung.com>